[model-gateway] extract conversation out of oai router
