+++ sgl-router/src/routers/conversations/mod.rs
@@ -1,0 +1,8 @@
+//! Shared conversation management module.
+//!
+//! This module provides conversation CRUD operations that can be shared
+//! across different router implementations.
+
+mod handlers;
+
+pub use handlers::*;
