+++ sgl-router/src/server.rs
@@ -1,4 +1,5 @@
 use std::{
+    collections::HashMap,
     sync::{
         atomic::{AtomicBool, Ordering},
         Arc,
@@ -322,16 +323,22 @@ struct ListItemsQuery {
 async fn v1_conversations_list_items(
     State(state): State<Arc<AppState>>,
     Path(conversation_id): Path<String>,
+    headers: http::HeaderMap,
     Query(ListItemsQuery {
         limit,
         order,
         after,
     }): Query<ListItemsQuery>,
-    headers: http::HeaderMap,
 ) -> Response {
     state
         .router
-        .list_conversation_items(Some(&headers), &conversation_id, limit, order, after)
+        .list_conversation_items(
+            Some(&headers),
+            &conversation_id,
+            limit,
+            order.as_deref(),
+            after.as_deref(),
+        )
         .await
 }
 
@@ -356,8 +363,8 @@ async fn v1_conversations_create_items(
 async fn v1_conversations_get_item(
     State(state): State<Arc<AppState>>,
     Path((conversation_id, item_id)): Path<(String, String)>,
-    Query(query): Query<GetItemQuery>,
     headers: http::HeaderMap,
+    Query(query): Query<GetItemQuery>,
 ) -> Response {
     state
         .router
@@ -545,7 +552,7 @@ async fn get_worker(State(state): State<Arc<AppState>>, Path(url): Path<String>)
             tool_parser: None,
             chat_template: None,
             bootstrap_port: None,
-            metadata: std::collections::HashMap::new(),
+            metadata: HashMap::new(),
             job_status: Some(status),
         };
         return Json(worker_info).into_response();
